Experimenting with slash commands

+++ events/ready.js
@@ -56,16 +56,19 @@ const deleteExpiredSavedRollCommands = async () => {
   }
 }
 
-const removeWarningInteractivity = async (client, warning) => {
+const removeWarningInteractivity = async (warning) => {
   let channel
   try {
-    channel = await client.channels.fetch(warning[MESSAGES_COLUMNS.channel_id])
+    channel = await Client.getChannelById(warning[MESSAGES_COLUMNS.channel_id])
   } catch (error) {
     if (error && error.name === 'DiscordAPIError') {
       logger.warn(`Failed to fetch channel for a warning`, error)
     } else {
       logger.error(`Failed to fetch channel for a warning`, error)
     }
+    return
+  }
+  if (!channel) {
     return
   }
   let message
@@ -79,7 +82,9 @@ const removeWarningInteractivity = async (client, warning) => {
     }
     return
   }
-
+  if (!message) {
+    return
+  }
   try {
     await message.delete()
   } catch (error) {
@@ -111,7 +116,7 @@ const deleteExpiredWarningMessages = async () => {
         const channelId = warning[MESSAGES_COLUMNS.channel_id]
         const messageId = warning[MESSAGES_COLUMNS.message_id]
         Client.reactionsCache[channelId+ '_' + messageId] = null
-        removeWarningInteractivity(Client.client, warning)
+        removeWarningInteractivity(warning)
       })
     }
   } catch (error) {
@@ -171,7 +176,7 @@ module.exports = async (client) => {
   const DBL = require('dblapi.js')
   const dbl = new DBL(process.env.DBL_TOKEN, client)
 
-  if(client.shard) {
+  if (client.shard) {
     logger.log(`Shard ids: "${client.shard.ids.join('", "')}"; count: "${client.shard.count}"`)
   }
   else {
@@ -190,6 +195,17 @@ module.exports = async (client) => {
       }
     }
   }, transformMinutesToMs(30))
+
+  try {
+    const channel = await Client.getChannelById(process.env.LOG_CHANNEL_ID)
+    if (channel) {
+      const mes = await Client.getMessageByIdAndChannelId(
+        channel.lastMessageId, channel.id)
+      let i = 0
+    }
+  } catch (err) {
+    logger.error(`Failed to get channel`, err)
+  }
 
   await tryToSetActivity()
   setInterval(async () => {
